added post controler and views

+++ app/controllers/users_controller.rb
@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
 
-	# before_filter :authorize
+	before_filter :authorize, except: [:new, :create]
 
 	def index
 		@users = User.all
@@ -8,40 +8,47 @@ class UsersController < ApplicationController
 	end
 
 	def new
-		render :new
+		if current_user
+			redirect_to posts_path(current_user)
+		else
+			render :new
+		end
 	end
 
 	def create
-		user_params = params.require(:user).permit(:email, :password)
-		user = User.new(user_params)
-		if user.save
-			session[:user_id] = user.id
-			redirect_to users_path(user)
+		if current_user
+			redirect_to posts_path(current_user)
 		else
-			flash[:error] = user.errors.full_messages.join(", ")	
-      		redirect_to new_user_path
+			user_params = params.require(:user).permit(:email, :password)
+			user = User.new(user_params)
+			if user.save
+				session[:user_id] = user.id
+				redirect_to posts_path
+			else
+				flash[:error] = user.errors.full_messages.join(", ")	
+	      		redirect_to new_user_path
+			end
 		end
 	end
 
 	def show
-		user_id = param[:id]
+		user_id = params[:id]
 		@user = User.find(user_id)
 		render :show
 	end
 
 	def edit
-		user_id = param[:id]
+		user_id = params[:id]
 		@user = User.find(user_id)
 		render :edit
 	end
 
 	def update
-		user_id = param[:id]
+		user_id = params[:id]
 		user = User.find(user_id)
 		user_params = params.require(:user).permit(:email, :password)
-		user.update_attribues(user_params)
-		if user.update
-			redirect_to user_path(user)
+		if user.update_attributes(user_params)
+			redirect_to posts_path(user)
 		else
 			flash[:error] = user.errors.full_messages.join(", ")
 			redirect_to edit_user_path(user)
@@ -51,13 +58,15 @@ class UsersController < ApplicationController
 	def destroy
 		user_id = params[:id]
 		user = User.find(user_id)
-		user.destroy
+		if current_user == user
+			session[:user_id] = nil
+			flash[:notice]="success"
+			user.destroy
+		
+		else
+			flash[:error]="YOU CAN'T"
+		end
 		redirect_to users_path
 	end
 
-	# private
-
-	#   def user_params
-	    
-	#   end
 end
